backend/requirements.lock.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.1
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.1.1
fastapi==0.110.0
greenlet==3.3.0
gunicorn==21.2.0
h11==0.16.0
httptools==0.7.1
idna==3.11
packaging==25.0
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
requests==2.31.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.25
starlette==0.36.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.27.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
PyJWT==2.8.0
